Playlists mostly working
Creating playlists & playlist page with playlist values working. Need to
create form to add audio to playlists.

Also need to go back and do the login & sign-up properly. Have the code
for this workign in another file, just need to transfer it over.

Need to add sessions & logout and change database to pool.